+++ CHANGES.md
@@ -17,8 +17,10 @@
 
 - Viewer role for team members  [Taiga #1056 & #6590](https://tree.taiga.io/project/penpot/us/1056 & https://tree.taiga.io/project/penpot/us/6590)
 - File history versions management [Taiga](https://tree.taiga.io/project/penpot/us/187?milestone=411120)
+- Rename selected layer via keyboard shortcut and context menu option [Taiga #8882](https://tree.taiga.io/project/penpot/us/8882)
 
 ### :bug: Bugs fixed
+
 
 ## 2.3.1